requirements-fastapi.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic[email]
transformers==4.35.2
torch>=2.1.0
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
#